Extract AddColumnTask from TransportAddColumnAction
Mostly to make it a bit easier to write unit tests for it
